backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
boto3>=1.35.0
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
